refactor(auth/pres): decouple register endpoint from app layer
- Replace RegisterCommand with RegisterRequest DTO as endpoint param
- Create Contracts.Register namespace to store DTO by logical use
- Move ErrorResponse in Contracts namespace
