+++ services/auth/src/Auth.Presentation/Contracts/Register/RegisterRequest.cs
@@ -1,0 +1,6 @@
+namespace Auth.Presentation.Contracts.Register;
+
+public sealed record RegisterRequest(
+    string Email,
+    string Password
+);
