+++ services/auth/src/Auth.Presentation/Contracts/Register/RegisterResponse.cs
@@ -1,4 +1,4 @@
-namespace Auth.Presentation.Contracts.Responses;
+namespace Auth.Presentation.Contracts.Register;
 
 public sealed record RegisterResponse(
     long UserId,
