--- services/auth/src/Auth.Presentation/Contracts/Responses/ErrorResponse.cs
@@ -1,3 +1,0 @@
-namespace Auth.Presentation.Contracts.Responses;
-
-public sealed record ErrorResponse(string Error);
